# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BlueHazy FM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shows:
    get:
      operationId: listShows
      tags: [shows]
      summary: List all shows
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of shows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Show"
    post:
      operationId: createShow
      tags: [shows]
      summary: Create a show
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Show"
  /shows/{id}:
    get:
      operationId: getShow
      tags: [shows]
      summary: Get a show
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Show
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Show"
    patch:
      operationId: updateShow
      tags: [shows]
      summary: Update a show
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShowUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Show"
    delete:
      operationId: deleteShow
      tags: [shows]
      summary: Delete a show
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Show:
      type: object
      required: [id, title, description, category, imageUrl, isFeatured, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        presenterId:
          type: ["integer", "null"]
        presenterName:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        createdAt:
          type: string
    ShowInput:
      type: object
      required: [title, description, category]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        presenterId:
          type: integer
        isFeatured:
          type: boolean
    ShowUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        presenterId:
          type: integer
        isFeatured:
          type: boolean
...