# lib/api-zod/src/generated/types/show.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BlueHazy FM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Show {
  id: number;
  title: string;
  description: string;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  presenterId?: number | null;
  /** @nullable */
  presenterName?: string | null;
  isFeatured: boolean;
  createdAt: string;
}